Example REBL Survey Data

A processed dataset containing simulated responses to the REBL survey from 100 participants. This dataset contains numeric (0/1) coded responses for 24 REBL items measuring pro-environmental behaviors across four domains: food, home, packaging, social, and water.

Usage

example

Format

A data frame with 100 rows and 25 variables:

respondent_id

Character. Unique participant identifier (p1-p100)

foodLocal

Numeric. Eating locally grown food (0=No, 1=Yes)

foodLunchNoMeat

Numeric. Eating lunch without meat (0=No, 1=Yes)

foodMeat

Numeric. Eating meat regularly (0=No, 1=Yes)

foodOatMilk

Numeric. Drinking oat milk (0=No, 1=Yes)

foodTofu

Numeric. Eating tofu (0=No, 1=Yes)

foodVegan

Numeric. Following vegan diet (0=No, 1=Yes)

homeClothesCold

Numeric. Washing clothes in cold water (0=No, 1=Yes)

homeClothesHang

Numeric. Hanging clothes to dry (0=No, 1=Yes)

homeLightsOff

Numeric. Turning lights off when leaving room (0=No, 1=Yes)

packCarriedUtensils

Numeric. Carrying reusable utensils (0=No, 1=Yes)

packCompost

Numeric. Composting organic waste (0=No, 1=Yes)

packContainerToRestaurant

Numeric. Bringing container to restaurant (0=No, 1=Yes)

packPickedUpLitter

Numeric. Picking up litter (0=No, 1=Yes)

packPullRecycleFromTrash

Numeric. Pulling recyclables from trash (0=No, 1=Yes)

packRags

Numeric. Using rags instead of paper towels (0=No, 1=Yes)

packReusableMug

Numeric. Using reusable mug (0=No, 1=Yes)

packReusedPaperPlasticBags

Numeric. Reusing paper/plastic bags (0=No, 1=Yes)

purchBuyNothing

Numeric. Participating in buy nothing groups (0=No, 1=Yes)

socialDocumentary

Numeric. Watching environmental documentaries (0=No, 1=Yes)

socialGroup

Numeric. Participating in environmental groups (0=No, 1=Yes)

socialRead

Numeric. Reading about environmental issues (0=No, 1=Yes)

socialSupportive

Numeric. Being supportive of environmental causes (0=No, 1=Yes)

waterShowerStop

Numeric. Stopping water while showering (0=No, 1=Yes)

waterTeethStop

Numeric. Stopping water while brushing teeth (0=No, 1=Yes)

Details

This dataset is the processed version of raw_example, where character responses ("Yes"/"No") have been converted to numeric values (1/0). The data can be used for Rasch modeling and other psychometric analyses of pro-environmental behavior patterns.

See also

Examples

# Load the data
data(example)

# View structure
str(example)
#> 'data.frame':	100 obs. of  25 variables:
#>  $ respondent_id             : chr  "p1" "p2" "p3" "p4" ...
#>  $ foodLocal                 : num  1 1 0 1 1 0 1 0 0 1 ...
#>  $ foodLunchNoMeat           : num  0 0 1 1 1 0 0 0 1 0 ...
#>  $ foodMeat                  : num  1 0 1 0 0 1 1 0 0 0 ...
#>  $ foodOatMilk               : num  0 0 0 1 1 1 0 0 0 0 ...
#>  $ foodTofu                  : num  0 0 1 1 1 1 1 1 0 0 ...
#>  $ foodVegan                 : num  0 0 0 1 0 1 0 0 1 0 ...
#>  $ homeClothesCold           : num  1 1 1 1 1 1 1 0 1 1 ...
#>  $ homeClothesHang           : num  1 0 0 0 1 1 0 0 0 0 ...
#>  $ homeLightsOff             : num  0 0 1 1 1 1 1 1 1 1 ...
#>  $ packCarriedUtensils       : num  0 0 0 1 0 1 0 0 0 0 ...
#>  $ packCompost               : num  1 0 1 0 0 0 0 1 0 0 ...
#>  $ packContainerToRestaurant : num  0 1 1 0 0 1 1 0 1 1 ...
#>  $ packPickedUpLitter        : num  1 1 1 1 0 0 1 0 0 1 ...
#>  $ packPullRecycleFromTrash  : num  1 1 0 0 0 1 1 1 1 0 ...
#>  $ packRags                  : num  0 1 1 1 1 1 1 1 1 1 ...
#>  $ packReusableMug           : num  0 1 0 0 0 0 1 0 0 0 ...
#>  $ packReusedPaperPlasticBags: num  0 1 0 1 1 0 1 0 1 1 ...
#>  $ purchBuyNothing           : num  0 0 1 0 0 1 0 0 1 1 ...
#>  $ socialDocumentary         : num  1 0 0 0 0 1 0 0 1 1 ...
#>  $ socialGroup               : num  1 1 1 1 0 1 0 0 0 1 ...
#>  $ socialRead                : num  1 0 0 0 1 1 1 1 1 1 ...
#>  $ socialSupportive          : num  0 0 1 1 0 0 1 1 1 0 ...
#>  $ waterShowerStop           : num  0 1 0 0 0 1 0 0 1 0 ...
#>  $ waterTeethStop            : num  1 1 0 1 0 0 0 1 0 0 ...

# Summary of pro-environmental behavior frequencies
colSums(example[, -1])  # Exclude respondent_id column
#>                  foodLocal            foodLunchNoMeat 
#>                         50                         52 
#>                   foodMeat                foodOatMilk 
#>                         28                         42 
#>                   foodTofu                  foodVegan 
#>                         42                         38 
#>            homeClothesCold            homeClothesHang 
#>                         63                         37 
#>              homeLightsOff        packCarriedUtensils 
#>                         57                         43 
#>                packCompost  packContainerToRestaurant 
#>                         38                         39 
#>         packPickedUpLitter   packPullRecycleFromTrash 
#>                         48                         35 
#>                   packRags            packReusableMug 
#>                         57                         44 
#> packReusedPaperPlasticBags            purchBuyNothing 
#>                         61                         36 
#>          socialDocumentary                socialGroup 
#>                         46                         50 
#>                 socialRead           socialSupportive 
#>                         48                         41 
#>            waterShowerStop             waterTeethStop 
#>                         46                         51