Raw REBL Survey Data

The unprocessed dataset containing simulated responses to the REBL survey from 100 participants. This dataset contains character responses ("Yes"/"No") for 24 REBL items measuring pro-environmental behaviors across four domains: food, home, packaging, social, and water.

Usage

raw_example

Format

A data frame with 100 rows and 25 variables:

respondent_id

Character. Unique participant identifier (p1-p100)

foodLocal

Character. Eating locally grown food ("Yes"/"No")

foodLunchNoMeat

Character. Eating lunch without meat ("Yes"/"No")

foodMeat

Character. Eating meat regularly ("Yes"/"No")

foodOatMilk

Character. Drinking oat milk ("Yes"/"No")

foodTofu

Character. Eating tofu ("Yes"/"No")

foodVegan

Character. Following vegan diet ("Yes"/"No")

homeClothesCold

Character. Washing clothes in cold water ("Yes"/"No")

homeClothesHang

Character. Hanging clothes to dry ("Yes"/"No")

homeLightsOff

Character. Turning lights off when leaving room ("Yes"/"No")

packCarriedUtensils

Character. Carrying reusable utensils ("Yes"/"No")

packCompost

Character. Composting organic waste ("Yes"/"No")

packContainerToRestaurant

Character. Bringing container to restaurant ("Yes"/"No")

packPickedUpLitter

Character. Picking up litter ("Yes"/"No")

packPullRecycleFromTrash

Character. Pulling recyclables from trash ("Yes"/"No")

packRags

Character. Using rags instead of paper towels ("Yes"/"No")

packReusableMug

Character. Using reusable mug ("Yes"/"No")

packReusedPaperPlasticBags

Character. Reusing paper/plastic bags ("Yes"/"No")

purchBuyNothing

Character. Participating in buy nothing groups ("Yes"/"No")

socialDocumentary

Character. Watching environmental documentaries ("Yes"/"No")

socialGroup

Character. Participating in environmental groups ("Yes"/"No")

socialRead

Character. Reading about environmental issues ("Yes"/"No")

socialSupportive

Character. Being supportive of environmental causes ("Yes"/"No")

waterShowerStop

Character. Stopping water while showering ("Yes"/"No")

waterTeethStop

Character. Stopping water while brushing teeth ("Yes"/"No")

Details

This dataset contains the original character-coded responses from the REBL survey before any processing. It represents how participants actually responded to each survey item. This data can be processed using functions like recode_rebl to convert to numeric format for analysis.

See also

Examples

# Load the data
data(raw_example)

# View structure
str(raw_example)
#> 'data.frame':	100 obs. of  25 variables:
#>  $ respondent_id             : chr  "p1" "p2" "p3" "p4" ...
#>  $ foodLocal                 : chr  "Yes" "Yes" "No" "Yes" ...
#>  $ foodLunchNoMeat           : chr  "No" "No" "Yes" "Yes" ...
#>  $ foodMeat                  : chr  "Yes" "No" "Yes" "No" ...
#>  $ foodOatMilk               : chr  "No" "No" "No" "Yes" ...
#>  $ foodTofu                  : chr  "No" "No" "Yes" "Yes" ...
#>  $ foodVegan                 : chr  "No" "No" "No" "Yes" ...
#>  $ homeClothesCold           : chr  "Yes" "Yes" "Yes" "Yes" ...
#>  $ homeClothesHang           : chr  "Yes" "No" "No" "No" ...
#>  $ homeLightsOff             : chr  "No" "No" "Yes" "Yes" ...
#>  $ packCarriedUtensils       : chr  "No" "No" "No" "Yes" ...
#>  $ packCompost               : chr  "Yes" "No" "Yes" "No" ...
#>  $ packContainerToRestaurant : chr  "No" "Yes" "Yes" "No" ...
#>  $ packPickedUpLitter        : chr  "Yes" "Yes" "Yes" "Yes" ...
#>  $ packPullRecycleFromTrash  : chr  "Yes" "Yes" "No" "No" ...
#>  $ packRags                  : chr  "No" "Yes" "Yes" "Yes" ...
#>  $ packReusableMug           : chr  "No" "Yes" "No" "No" ...
#>  $ packReusedPaperPlasticBags: chr  "No" "Yes" "No" "Yes" ...
#>  $ purchBuyNothing           : chr  "No" "No" "Yes" "No" ...
#>  $ socialDocumentary         : chr  "Yes" "No" "No" "No" ...
#>  $ socialGroup               : chr  "Yes" "Yes" "Yes" "Yes" ...
#>  $ socialRead                : chr  "Yes" "No" "No" "No" ...
#>  $ socialSupportive          : chr  "No" "No" "Yes" "Yes" ...
#>  $ waterShowerStop           : chr  "No" "Yes" "No" "No" ...
#>  $ waterTeethStop            : chr  "Yes" "Yes" "No" "Yes" ...

# Count "Yes" responses for each item
sapply(raw_example[, -1], function(x) sum(x == "Yes"))
#>                  foodLocal            foodLunchNoMeat 
#>                         50                         52 
#>                   foodMeat                foodOatMilk 
#>                         28                         42 
#>                   foodTofu                  foodVegan 
#>                         42                         38 
#>            homeClothesCold            homeClothesHang 
#>                         63                         37 
#>              homeLightsOff        packCarriedUtensils 
#>                         57                         43 
#>                packCompost  packContainerToRestaurant 
#>                         38                         39 
#>         packPickedUpLitter   packPullRecycleFromTrash 
#>                         48                         35 
#>                   packRags            packReusableMug 
#>                         57                         44 
#> packReusedPaperPlasticBags            purchBuyNothing 
#>                         61                         36 
#>          socialDocumentary                socialGroup 
#>                         46                         50 
#>                 socialRead           socialSupportive 
#>                         48                         41 
#>            waterShowerStop             waterTeethStop 
#>                         46                         51 

# Convert to numeric format
# processed_data <- recode_rebl(raw_example)